refactor: moves asyncWrap function
